refactor: [code-review] Fix the transitive `runtime -> bootstrap -> app… [ORB-11631]
[code-review] Fix the transitive `runtime -> bootstrap -> application` import and close the guardrail gap

Planned-By: codex
Orchestrated-By: astra
