+++ crates/orbit-core/src/application/workspace_sync.rs
@@ -15,7 +15,7 @@ use super::{
     ManagedAssetReconciliation, reconcile_managed_assets_in_mode,
 };
 use crate::application::auto_tasks::{DEFAULT_AUTO_TASK_FILES, auto_tasks_dir};
-use crate::bootstrap::activity::DEFAULT_ACTIVITY_FILES;
+use crate::runtime::assets::DEFAULT_ACTIVITY_FILES;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize)]
 #[serde(rename_all = "snake_case")]
