+++ crates/orbit-core/src/runtime/assets.rs
@@ -1,0 +1,189 @@
+//! Embedded activity definitions available to the runtime and bootstrap layers.
+
+/// Shippable default activity assets, seeded under
+/// `<orbit_root>/resources/activities/<name>.yaml` on `orbit init`. Keep this
+/// list in sync with the workflow YAMLs under `crates/orbit-core/assets/jobs/`:
+/// every `target: activity:<name>` reference in a shipped workflow must
+/// resolve to an entry here. Reference/example activities (anything under
+/// `assets/activities/examples/`) are deliberately excluded — they're
+/// fixtures for `crates/orbit-engine/examples/v2_job_runtime_smoke.rs`, not
+/// runtime defaults.
+pub(crate) const DEFAULT_ACTIVITY_FILES: &[(&str, &str)] = &[
+    (
+        "agent_implement",
+        include_str!("../../assets/activities/agent_implement.yaml"),
+    ),
+    (
+        "apply_triage_dispositions",
+        include_str!("../../assets/activities/apply_triage_dispositions.yaml"),
+    ),
+    (
+        "apply_task_pilot_results",
+        include_str!("../../assets/activities/apply_task_pilot_results.yaml"),
+    ),
+    (
+        "classify_workspace_auto_tasks",
+        include_str!("../../assets/activities/classify_workspace_auto_tasks.yaml"),
+    ),
+    (
+        "collect_ci_evidence",
+        include_str!("../../assets/activities/collect_ci_evidence.yaml"),
+    ),
+    (
+        "collect_dependabot_alerts",
+        include_str!("../../assets/activities/collect_dependabot_alerts.yaml"),
+    ),
+    (
+        "drain_window",
+        include_str!("../../assets/activities/drain_window.yaml"),
+    ),
+    (
+        "file_ci_failure_tasks",
+        include_str!("../../assets/activities/file_ci_failure_tasks.yaml"),
+    ),
+    (
+        "file_dependabot_alert_tasks",
+        include_str!("../../assets/activities/file_dependabot_alert_tasks.yaml"),
+    ),
+    (
+        "epic_orchestrator",
+        include_str!("../../assets/activities/epic_orchestrator.yaml"),
+    ),
+    (
+        "agent_invoke",
+        include_str!("../../assets/activities/agent_invoke.yaml"),
+    ),
+    (
+        "gate_starvation_fail",
+        include_str!("../../assets/activities/gate_starvation_fail.yaml"),
+    ),
+    (
+        "git_merge",
+        include_str!("../../assets/activities/git_merge.yaml"),
+    ),
+    (
+        "git_commit",
+        include_str!("../../assets/activities/git_commit.yaml"),
+    ),
+    (
+        "git_push",
+        include_str!("../../assets/activities/git_push.yaml"),
+    ),
+    (
+        "git_rebase",
+        include_str!("../../assets/activities/git_rebase.yaml"),
+    ),
+    (
+        "invoke_and_wait",
+        include_str!("../../assets/activities/invoke_and_wait.yaml"),
+    ),
+    (
+        "invoke_detached",
+        include_str!("../../assets/activities/invoke_detached.yaml"),
+    ),
+    (
+        "pipeline_success_guard",
+        include_str!("../../assets/activities/pipeline_success_guard.yaml"),
+    ),
+    (
+        "list_backlog_tasks",
+        include_str!("../../assets/activities/list_backlog_tasks.yaml"),
+    ),
+    (
+        "list_triage_candidates",
+        include_str!("../../assets/activities/list_triage_candidates.yaml"),
+    ),
+    (
+        "scan_unresolved_work",
+        include_str!("../../assets/activities/scan_unresolved_work.yaml"),
+    ),
+    (
+        "pr_open",
+        include_str!("../../assets/activities/pr_open.yaml"),
+    ),
+    (
+        "pr_failure_handoff",
+        include_str!("../../assets/activities/pr_failure_handoff.yaml"),
+    ),
+    (
+        "pr_conflict_recovery",
+        include_str!("../../assets/activities/pr_conflict_recovery.yaml"),
+    ),
+    (
+        "pr_prepare",
+        include_str!("../../assets/activities/pr_prepare.yaml"),
+    ),
+    (
+        "pr_complete",
+        include_str!("../../assets/activities/pr_complete.yaml"),
+    ),
+    (
+        "pr_promote",
+        include_str!("../../assets/activities/pr_promote.yaml"),
+    ),
+    (
+        "review_gate_admit",
+        include_str!("../../assets/activities/review_gate_admit.yaml"),
+    ),
+    (
+        "agent_review_repair",
+        include_str!("../../assets/activities/agent_review_repair.yaml"),
+    ),
+    (
+        "review_gate_settle",
+        include_str!("../../assets/activities/review_gate_settle.yaml"),
+    ),
+    (
+        "prepare_task_pilot",
+        include_str!("../../assets/activities/prepare_task_pilot.yaml"),
+    ),
+    (
+        "reserve_locks",
+        include_str!("../../assets/activities/reserve_locks.yaml"),
+    ),
+    (
+        "release_locks",
+        include_str!("../../assets/activities/release_locks.yaml"),
+    ),
+    (
+        "resolve_workspace_ship_input",
+        include_str!("../../assets/activities/resolve_workspace_ship_input.yaml"),
+    ),
+    (
+        "run_auto_task_scheduler",
+        include_str!("../../assets/activities/run_auto_task_scheduler.yaml"),
+    ),
+    ("sleep", include_str!("../../assets/activities/sleep.yaml")),
+    (
+        "step_failure_recovery",
+        include_str!("../../assets/activities/step_failure_recovery.yaml"),
+    ),
+    (
+        "task_complete",
+        include_str!("../../assets/activities/task_complete.yaml"),
+    ),
+    (
+        "task_pilot",
+        include_str!("../../assets/activities/task_pilot.yaml"),
+    ),
+    (
+        "triage_failed_runs",
+        include_str!("../../assets/activities/triage_failed_runs.yaml"),
+    ),
+    (
+        "update_task",
+        include_str!("../../assets/activities/update_task.yaml"),
+    ),
+    (
+        "validate_bundles",
+        include_str!("../../assets/activities/validate_bundles.yaml"),
+    ),
+    (
+        "worktree_setup",
+        include_str!("../../assets/activities/worktree_setup.yaml"),
+    ),
+    (
+        "worktree_gc",
+        include_str!("../../assets/activities/worktree_gc.yaml"),
+    ),
+];
